requirements.txt
Django>=4.2
Pillow>=10.0
torch>=2.1
torchaudio>=2.1
transformers>=4.38
vosk>=0.3.45
soundfile>=0.12.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5
